Index doc type for stats announcments
This will add content_store_document_type and public_updated_at
to stats announcements in rummager, and will add release_timestamp
to published stats in rummager.

This is useful because it lets us sort on both time fields when
looking at statistics and their corresponding announcements.
The document type is useful as it lets us query announcements by
document type which is consistent with other document types.

We are going to run rake rummager:index:government on integration
to try this out.