Remove org page rendering
Courts and tribunals pages have now been migrated to Collections so we can
remove most of the organisations controller code.

The courts and tribunals index page is still rendered by Whitehall, so
we need to keep the index action around for a bit.

The only atom feeds that we were serving were for courts and tribunals
pages, but none of these had any content, so 404ing them is probably OK.
It'd be more of a shock if we redirected them to a "non-Whitehall" feed
for the courts and tribunals, because then they'd get a torrent of content
that is unexpected.

A lot of tests are going to fail now because lots of things check defunct
org pages to see if they've worked.