Filter out locked documents
We don't want users to be able to see locked documents when they are choosing
documents to feature so we adapt the Edition Filter to default to filtering
locked documents out. The exception to this is the documents index page as we
still want users to be able to see locked documents in the list.