Remove the organisations api
This was migrated to [the collections app](https://github.com/alphagov/collections/blob/master/app/controllers/organisations_api_controller.rb)
ages ago.  I've verified that there are no requests coming to this controller
by running a query for `application: whitehall AND controller: api*` in
Kibana and verifying that the only requests are for the following controllers:

- api/governments
- api/world_locations
- api/worldwide_organisations
- api/user

It's fine to remove references to this from organisation pages in this application,
because they are also now rendered by collections!