Fix contacts private method delegation warning
Using def_delegators on the translatable methods of a globalize enhanced
ActiveRecord model is at risk of causing warnings. This is because, as
far as I can tell, a reliance on meta programming and method missing for
method existence. This seems to kick in when the translatable record is
updated (presumably as it may need to reconfigure the methods of the
model). The forwardable module tries to determine whether a
method exists publicly or not before calling otherwise it raises a
warning.